pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "flowfeed"
version = "1.0.0"
description = "Lightweight intelligent news aggregation engine for developers — CLI-first, multi-source, multi-format export"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "gitstq", email = "gitstq@users.noreply.github.com"}
]
keywords = ["news", "aggregation", "cli", "trending", "hackernews", "rss", "developer-tools"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Utilities",
]
dependencies = [
    "httpx>=0.27.0",
    "rich>=13.0.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
scheduler = ["apscheduler>=3.10.0"]
mcp = ["mcp>=1.0.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
]

[project.scripts]
flowfeed = "flowfeed.cli:main"

[project.urls]
Homepage = "https://github.com/gitstq/flowfeed"
Repository = "https://github.com/gitstq/flowfeed"
Issues = "https://github.com/gitstq/flowfeed/issues"

[tool.setuptools.packages.find]
include = ["flowfeed*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
